Fix bug with bamtools reporting incorrect tags

+++ src/utils.cpp
@@ -241,13 +241,13 @@ string convertToSam(const BamAlignment& a, RefVector m_references) {
 
             case (Constants::BAM_TAG_TYPE_INT8):
                 // force value into integer-type (instead of char value)
-                m_out << "i:" << static_cast<int16_t>(tagData[index]);
+                m_out << "i:" << int(static_cast<int8_t>(tagData[index]));
                 ++index;
                 break;
 
             case (Constants::BAM_TAG_TYPE_UINT8):
                 // force value into integer-type (instead of char value)
-                m_out << "i:" << static_cast<uint16_t>(tagData[index]);
+                m_out << "i:" << int(static_cast<uint8_t>(tagData[index]));
                 ++index;
                 break;
